Move `center_times()` to `axis.py` - Update references to `center_times()` - Add function `_get_coord_var()` - Refactor tests in `test_axis.py`
