+++ tests/test_axis.py
@@ -3,11 +3,99 @@ import pytest
 import xarray as xr
 
 from tests.fixtures import generate_dataset
-from xcdat.axis import (
-    _align_lon_bounds_to_360,
-    _get_prime_meridian_index,
-    swap_lon_axis,
-)
+from xcdat.axis import center_times, swap_lon_axis
+
+
+class TestCenterTimes:
+    @pytest.fixture(autouse=True)
+    def setup(self):
+        self.ds = generate_dataset(cf_compliant=True, has_bounds=True)
+
+    def test_raises_error_if_time_coord_var_does_not_exist_in_dataset(self):
+        ds = self.ds.copy()
+        ds = ds.drop_dims("time")
+
+        with pytest.raises(KeyError):
+            center_times(ds)
+
+    def test_raises_error_if_time_bounds_does_not_exist_in_the_dataset(self):
+        ds = self.ds.copy()
+        ds = ds.drop_vars("time_bnds")
+
+        with pytest.raises(KeyError):
+            center_times(ds)
+
+    def test_gets_time_as_the_midpoint_between_time_bounds(self):
+        ds = self.ds.copy()
+
+        # Make the time coordinates uncentered.
+        uncentered_time = np.array(
+            [
+                "2000-01-31T12:00:00.000000000",
+                "2000-02-29T12:00:00.000000000",
+                "2000-03-31T12:00:00.000000000",
+                "2000-04-30T00:00:00.000000000",
+                "2000-05-31T12:00:00.000000000",
+                "2000-06-30T00:00:00.000000000",
+                "2000-07-31T12:00:00.000000000",
+                "2000-08-31T12:00:00.000000000",
+                "2000-09-30T00:00:00.000000000",
+                "2000-10-16T12:00:00.000000000",
+                "2000-11-30T00:00:00.000000000",
+                "2000-12-31T12:00:00.000000000",
+                "2001-01-31T12:00:00.000000000",
+                "2001-02-28T00:00:00.000000000",
+                "2001-12-31T12:00:00.000000000",
+            ],
+            dtype="datetime64[ns]",
+        )
+        ds.time.data[:] = uncentered_time
+
+        # Compare result of the method against the expected.
+        expected = ds.copy()
+        expected_time_data = np.array(
+            [
+                "2000-01-16T12:00:00.000000000",
+                "2000-02-15T12:00:00.000000000",
+                "2000-03-16T12:00:00.000000000",
+                "2000-04-16T00:00:00.000000000",
+                "2000-05-16T12:00:00.000000000",
+                "2000-06-16T00:00:00.000000000",
+                "2000-07-16T12:00:00.000000000",
+                "2000-08-16T12:00:00.000000000",
+                "2000-09-16T00:00:00.000000000",
+                "2000-10-16T12:00:00.000000000",
+                "2000-11-16T00:00:00.000000000",
+                "2000-12-16T12:00:00.000000000",
+                "2001-01-16T12:00:00.000000000",
+                "2001-02-15T00:00:00.000000000",
+                "2001-12-16T12:00:00.000000000",
+            ],
+            dtype="datetime64[ns]",
+        )
+        expected = expected.assign_coords(
+            {
+                "time": xr.DataArray(
+                    name="time",
+                    data=expected_time_data,
+                    coords={"time": expected_time_data},
+                    dims="time",
+                    attrs={
+                        "long_name": "time",
+                        "standard_name": "time",
+                        "axis": "T",
+                        "bounds": "time_bnds",
+                    },
+                )
+            }
+        )
+        # Update time bounds with centered time coordinates.
+        time_bounds = ds.time_bnds.copy()
+        time_bounds["time"] = expected.time
+        expected["time_bnds"] = time_bounds
+
+        result = center_times(ds)
+        assert result.identical(expected)
 
 
 class TestSwapLonAxis:
@@ -16,7 +104,87 @@ class TestSwapLonAxis:
         with pytest.raises(ValueError):
             swap_lon_axis(ds, to=9000)  # type: ignore
 
-    def test_swap_from_180_to_360_and_sorts_with_prime_meridian_cell(self):
+    def test_raises_error_if_lon_bounds_contains_more_than_one_prime_meridian_cell(
+        self,
+    ):
+        ds_180 = xr.Dataset(
+            coords={
+                "lon": xr.DataArray(
+                    name="lon",
+                    data=np.array([-180, -1, 0, 1, 179]),
+                    dims=["lon"],
+                    attrs={"units": "degrees_east", "axis": "X", "bounds": "lon_bnds"},
+                )
+            },
+            data_vars={
+                "lon_bnds": xr.DataArray(
+                    name="lon_bnds",
+                    data=np.array(
+                        [
+                            [-180.5, -1.5],
+                            [-1.5, -0.5],
+                            [-0.5, 0.5],
+                            [0.5, 1.5],
+                            [-180.5, 1.5],
+                        ]
+                    ),
+                    dims=["lon", "bnds"],
+                    attrs={"is_generated": "True"},
+                ),
+                "ts": xr.DataArray(
+                    name="ts",
+                    data=np.array([0, 1, 2, 3, 4]),
+                    dims=["lon"],
+                    attrs={"test_attr": "test"},
+                ),
+            },
+        )
+
+        with pytest.raises(ValueError):
+            swap_lon_axis(ds_180, to=(0, 360))
+
+    def test_swap_from_360_to_180_and_sorts(self):
+        ds_360 = xr.Dataset(
+            coords={
+                "lon": xr.DataArray(
+                    name="lon",
+                    data=np.array([60, 150, 271]),
+                    dims=["lon"],
+                    attrs={"units": "degrees_east", "axis": "X", "bounds": "lon_bnds"},
+                )
+            },
+            data_vars={
+                "lon_bnds": xr.DataArray(
+                    name="lon_bnds",
+                    data=np.array([[0, 120], [120, 181], [181, 360]]),
+                    dims=["lon", "bnds"],
+                    attrs={"is_generated": "True"},
+                )
+            },
+        )
+
+        result = swap_lon_axis(ds_360, to=(-180, 180))
+        expected = xr.Dataset(
+            coords={
+                "lon": xr.DataArray(
+                    data=np.array([-89, 60, 150]),
+                    dims=["lon"],
+                    attrs={"units": "degrees_east", "axis": "X", "bounds": "lon_bnds"},
+                )
+            },
+            data_vars={
+                "lon_bnds": xr.DataArray(
+                    name="lon_bnds",
+                    data=np.array([[-179, 0], [0, 120], [120, -179]]),
+                    dims=["lon", "bnds"],
+                    attrs={"is_generated": "True"},
+                )
+            },
+        )
+
+        assert result.identical(expected)
+
+    def test_swap_from_180_to_360_and_sorts_with_prime_meridian_cell_in_lon_bnds(self):
         ds_180 = xr.Dataset(
             coords={
                 "lon": xr.DataArray(
@@ -86,140 +254,3 @@ class TestSwapLonAxis:
         )
 
         assert result.identical(expected)
-
-    def test_swap_from_360_to_180_and_sorts(self):
-        ds_360 = xr.Dataset(
-            coords={
-                "lon": xr.DataArray(
-                    name="lon",
-                    data=np.array([60, 150, 271]),
-                    dims=["lon"],
-                    attrs={"units": "degrees_east", "axis": "X", "bounds": "lon_bnds"},
-                )
-            },
-            data_vars={
-                "lon_bnds": xr.DataArray(
-                    name="lon_bnds",
-                    data=np.array([[0, 120], [120, 181], [181, 360]]),
-                    dims=["lon", "bnds"],
-                    attrs={"is_generated": "True"},
-                )
-            },
-        )
-
-        result = swap_lon_axis(ds_360, to=(-180, 180))
-        expected = xr.Dataset(
-            coords={
-                "lon": xr.DataArray(
-                    data=np.array([-89, 60, 150]),
-                    dims=["lon"],
-                    attrs={"units": "degrees_east", "axis": "X", "bounds": "lon_bnds"},
-                )
-            },
-            data_vars={
-                "lon_bnds": xr.DataArray(
-                    name="lon_bnds",
-                    data=np.array([[-179, 0], [0, 120], [120, -179]]),
-                    dims=["lon", "bnds"],
-                    attrs={"is_generated": "True"},
-                )
-            },
-        )
-
-        assert result.identical(expected)
-
-
-class TestAlignLonBoundsto360:
-    @pytest.fixture(autouse=True)
-    def setup(self):
-        self.ds = generate_dataset(cf_compliant=True, has_bounds=True)
-
-    def test_raises_error_if_bounds_below_0(self):
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            data=np.array([[-1, 1], [1, 90], [90, 180], [180, 359]]),
-            dims=["lon", "bnds"],
-        )
-        with pytest.raises(ValueError):
-            _align_lon_bounds_to_360(domain_bounds, np.array([0]))
-
-    def test_raises_error_if_bounds_above_360(self):
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            data=np.array([[359, 361], [1, 90], [90, 180], [180, 359]]),
-            dims=["lon", "bnds"],
-        )
-        with pytest.raises(ValueError):
-            _align_lon_bounds_to_360(domain_bounds, np.array([0]))
-
-    def test_extends_bounds_array_for_cell_spanning_prime_meridian(self):
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            coords={
-                "lon": xr.DataArray(
-                    name="lon",
-                    data=np.array([0, 90, 180, 359]),
-                    dims=["lon"],
-                    attrs={"axis": "X"},
-                )
-            },
-            data=np.array([[359, 1], [1, 90], [90, 180], [180, 359]]),
-            dims=["lon", "bnds"],
-        )
-
-        result_bounds = _align_lon_bounds_to_360(domain_bounds, np.array([0]))
-        expected_bounds = xr.DataArray(
-            name="lon_bnds",
-            coords={
-                "lon": xr.DataArray(
-                    name="lon",
-                    data=np.array([0, 90, 180, 359, 0]),
-                    dims=["lon"],
-                    attrs={"axis": "X"},
-                )
-            },
-            data=np.array([[0, 1], [1, 90], [90, 180], [180, 359], [359, 360]]),
-            dims=["lon", "bnds"],
-        )
-        assert result_bounds.identical(expected_bounds)
-
-    def test_retains_total_weight(self):
-        # construct array spanning 0 to 360
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            coords={
-                "lon": xr.DataArray(
-                    name="lon",
-                    data=np.array([0, 90, 180, 359]),
-                    dims=["lon"],
-                    attrs={"axis": "X"},
-                )
-            },
-            data=np.array([[359, 1], [1, 90], [90, 180], [180, 359]]),
-            dims=["lon", "bnds"],
-        )
-
-        result_bounds = _align_lon_bounds_to_360(domain_bounds, np.array(0))
-        dbdiff = np.sum(np.array(result_bounds[:, 1] - result_bounds[:, 0]))
-        assert dbdiff == 360.0
-
-
-class TestGetPrimeMeridianIndex:
-    def test_raises_error_if_multiple_bounds_span_prime_meridian(self):
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            data=np.array([[359, 1], [1, 90], [90, 180], [180, 2]]),
-            dims=["lon", "bnds"],
-        )
-        with pytest.raises(ValueError):
-            _get_prime_meridian_index(domain_bounds)
-
-    def test_returns_none_if_there_is_no_prime_meridian(self):
-        domain_bounds = xr.DataArray(
-            name="lon_bnds",
-            data=np.array([[0, 1], [1, 90], [90, 180], [180, 360]]),
-            dims=["lon", "bnds"],
-        )
-        result = _get_prime_meridian_index(domain_bounds)
-
-        assert result is None
